Remowe models in separate file. Split logic for different abstract class. Inherit Book class from separate classes. Used super() for acting currect method from parrent classes
